start rewriting parameter doc extraction
