+++ src/util.rs
@@ -1,5 +1,5 @@
 use quote::quote;
-use syn::{Attribute, Expr, LitStr, Meta, MetaNameValue};
+use syn::{Attribute, Expr, FnArg, Ident, LitStr, Meta, MetaNameValue, Pat};
 
 use crate::is_arguments_section;
 
@@ -59,6 +59,8 @@ pub struct FunctionDocs {
     pub after_args_section: Vec<Attribute>,
 }
 
+/// extract the documentation from the doc comments of the function and perform
+/// some additional logic
 pub fn extract_fn_doc_attrs(attrs: &mut Vec<Attribute>) -> Result<FunctionDocs, syn::Error> {
     let mut before_args_section = Vec::with_capacity(attrs.len());
     let mut after_args_section = Vec::with_capacity(attrs.len());
@@ -99,4 +101,47 @@ pub fn extract_fn_doc_attrs(attrs: &mut Vec<Attribute>) -> Result<FunctionDocs, 
         before_args_section,
         after_args_section,
     })
+}
+
+/// an identifier (such as a function parameter or a generic type)
+/// with doc attributes
+pub struct DocumentedIdent<'a> {
+    pub ident: &'a Ident,
+    /// the doc comments
+    pub docs: Vec<Attribute>,
+}
+
+impl<'a> DocumentedIdent<'a> {
+    pub fn new(ident: &'a Ident, docs: Vec<Attribute>) -> Self {
+        Self { ident, docs }
+    }
+}
+
+/// extract the parameter documentation from an iterator over function arguments.
+pub fn extract_documented_parameters<'a, I>(args: I) -> Result<Vec<DocumentedIdent<'a>>, syn::Error>
+where
+    I: Iterator<Item = &'a mut FnArg>,
+{
+    // will contain the docs comments for each documented function parameter
+    // together with the identifier of the function parameter.
+    let (lower, upper) = args.size_hint();
+    let mut documented_params = Vec::<DocumentedIdent>::with_capacity(upper.unwrap_or(lower));
+
+    for arg in args {
+        match arg {
+            FnArg::Typed(pat_type) => {
+                let Pat::Ident(pat_ident) = pat_type.pat.as_ref() else {
+                    unreachable!("unexpected node while parsing");
+                };
+                let ident = &pat_ident.ident;
+                let docs = extract_doc_attrs(&mut pat_type.attrs);
+
+                if !docs.is_empty() {
+                    documented_params.push(DocumentedIdent::new(ident, docs));
+                }
+            }
+            FnArg::Receiver(_) => {}
+        }
+    }
+    Ok(documented_params)
 }
